VSCodeWorkSpace/ProyectosPython/Proyecto texto a audio/requirements.txt
newspaper3k
gTTS